docs(fetch): clarify default fetch behavior (#72982)
## Why?

The default fetch behavior, which is shown as `auto no cache` when you
set the config
[logging](https://nextjs.org/docs/app/api-reference/next-config-js/logging)
to `true`, essentially means fetch once on `next build` to prerender the
page. But, when a [Dynamic
API](https://nextjs.org/docs/app/building-your-application/rendering/server-components#dynamic-rendering)
is used, it will fetch from the resource on every request.

This is different from `{ cache: 'no-store' }`, which regardless of
whether Dynamic APIs are used, will de-op the page from prerendering.