chore(CI): Remove sysroot override for macos arm64 (#73188)
This is extracted from #72513. This override appears to break the build for rocksdb. It doesn't seem like there's any good reason for us to do this.

It looks like we used to do a cross-build for macos arm64 (#28138), which would explain the need to override the sysroot, but we don't do that anymore. We don't do this override for x86-64 builds (even though that one _is_ a cross-build? The napi cli probably does the right thing for us...).

Manually triggered `build-and-deploy` CI job: https://github.com/vercel/next.js/actions/runs/12018322261