Turbopack: make client references actual references (#73174)
aka "one graph"

- Make client references actual references (non-chunkable, so chunking
ignores them right now. We can change that later and move more chunking
into Turbopack itself as opposed to the manual stuff in `AppEndpoint`)
- `find server actions` now just walks the rsc_entry and discovery
client references itself
- `collecting next/dynamic` still walks the RSC graph which is wrong (as
before). I left this to make this PR a noop that shouldn't change any
behavior. We'll fix this in the coming reimplementation (and add tests
for what is apparently currently not asserted 😐)

**View without whitespace changes**

Closes PACK-3532