[Segment Cache] Remove task map
Upon reconsideration, I don't think we need to store the tasks in a map.
I had thought we would use this to dedupe multiple prefetch tasks to the
same URL, but we already dedupe by URL when accessing the route entry.
In the future, there will also be multiple kinds of prefetch tasks, so
it will be totally possible for the same URL to have multiple tasks
associated with it; for example, a prefetch initiated by a <Link>
entering the viewport, versus a prefetch initiated by an imperative
prefetch() call. The <Link> prefetch should be cancelable, but the
imperative prefetch should not. So it just makes more sense for them to
be separate tasks.