refactor: remove isStandaloneMode context from metadata resolver (#73224)
This removes a check that appears to be redundant: we only want to warn about the fallback behavior when deployed in an environment where the env vars needed to detect proper fallback values (eg, not deployed to Vercel) aren't available. Since this check is already captured, we don't need both. 

The rationale for explicitly logging in standalone mode (even in dev) is that `standalone` mode won't ever be deployed to Vercel. But it's equally possible for you to build a Next.js app and not deploy it to Vercel. At best, this was reducing some additional console noise. But at worst, you're deploying an app that will serve a `localhost` URL in production. If you're using a relative URL without a metadataBase, it seems worth explicitly calling out as early as possible that we're going to fill in a fallback for you. 

Ultimately we should surface this information regardless of env in devtools. But since warning is the default, in #73066 I've updated it to only warn in dev when explicitly passing a relative og URL, rather than always making it warn when using the `opengraph-image` convention.