port turbopack-node to ResolvedVc (#73082)
Does what it says on the tin.

One struct seems to be trivially used in a trait which requires it to be a Vc so I left it with a comment. @bgw perhaps we can set up some lint + disabling comment for all the cases in this doc?

Would be handy to ensure we are continuing to follow the heuristic in the future also.