Collect function declarations that server functions close over
Instead of erroring with `ReferenceError: fn is not defined` when
triggering a server function, we can now show the following error early:

```
Error: Functions cannot be passed directly to Client Components unless
you explicitly expose it by marking it with "use server". Or maybe you
meant to call this function rather than return it.
```