Fix missing error message in terminal for bound args serialization error (#73508)
This is a small follow-up to #73471. The error message got lost when
assigning the custom stack (the message is usually also part of the
stack). This was only noticeable in the Terminal though; the dev overlay
showed the error correctly. We're now adding assertions on the CLI
output to the e2e test.

In addition, the custom stack slicing is replaced with
`Error.captureStackTrace` (h/t @eps1lon).