[Segment Cache] Add isPartial to segment prefetch  (#73528)
Based on

- #73434
- #73486 

---


During a navigation, we should be able to skip the dynamic request if
the prefetched data does not contain any dynamic holes. In the previous
cache implementation, we only tracked this per route; in the Segment
Cache, we must track this per segment.

This updates the SegmentCacheEntry and CacheNodeSeedData types to
include an `isPartial` field. The field is always false during a dynamic
render, or when PPR is disabled.

This PR does not change any behavior; it only adds the new field.