Fix missing client reference manifest error when using route groups (#73606)
While splitting up the `"use cache"` e2e test pages into two separate root layout groups (one with a suspense boundary, one without), I noticed that there's a bug when retrieving the client reference manifest for a page in a route group.

To fix it, we're now normalizing the `page` that's passed into `setReferenceManifestsSingleton` with the existing `normalizeAppPath` helper. When the manifest is retrieved, we're using the already normalized `workStore.route` instead of `workStore.page`.