fix: do not add suffix for sitemap under group routes (#73570)
### What

Do not add hash suffix for the metadata `/sitemap.xml` routes under the
routes group.

### Why

We previously added the suffix to all the metadata routes where they're
under the routes group. This was intentional for cases like having both
`/(group)/og image` and `/og image` so they can be different og images
inlined into different group routes. But for sitemap it's different
since sitemap usually wasn't intended to have duplicates. Each sitemap
contains all the urls of sub routes, we don't have the case
`/(group)/sitemap.[ext]` and `/sitemap.[ext]` exist at the same time. So
we always normalize the urls for sitemap and do not append hash suffix.

This was already the case in v14, it's regressed in #66477 

Fixes #72787
Closes NEXT-3909