[Turbopack] Introduce OperationVc that wraps operations (#70242)
### What?

Introduce `OperationVc` that wraps operations

*(Note: This was called `VcOperation`, but I changed it to `OperationVc` to better match the `ResolvedVc` naming convention -- @bgw)*

We should only expose `OperationVc` into JS to connect to the whole computation (and be strongly consistent with the whole computation).

~also fix query string~ Fixed in #70461

### Why?

We want operations to be strongly consistent to the whole operation. Also HMR subscriptions should include the whole entrypoints and endpoint operation. The `OperationVc` type makes it easy to track that and enforces connecting to the operation correctly.

In regards of the `ResolvedVc` work, we also want operations to be explicit.