Fix: Prefetching lazily generated param
Fixes an oversight in #72168 where the segment data was not correctly
transferred from the render result to the cache entry in the case where
a prerender is lazily generated after build.

I also took the opportunity to remove one of the intermediate types
used by the various layers that the segment data passes through. For
some reason, in the original PR I made the type of `segmentData` on
CachedAppPageValue an object while the corresponding type on
AppPageRenderResultMetadata was a map. (I didn't really notice before
because in the case where the entry is generated at built time, it gets
written to disk then read back out, so there's some data conversion
happening anyway.)